Breaking: use a separate script to publish to the npm registry
(refs https://github.com/eslint/eslint/issues/10631)

This updates the release API to not publish to npm when initially
called, and to only publish to npm when invoked as a separate process.
This will make it possible to pause the build to ask the user for a TOTP
code before publishing, provided that consumers of the package are
updated accordingly.