Update: avoid unnecessary checks for environment variables
Currently, eslint-release requires `NPM_TOKEN` and `ESLINT_GITHUB_TOKEN` environment variables to be present when generating a release in addition to publishing a release. However, the tokens aren't necessary when generating a release, and requiring them makes it more difficult to test-generate a release locally. With this commit, the tokens are only validated when publishing, not when generating.