Fix: avoid overwriting the .npmrc file in a repository
The release process needs to temporarily write an .npmrc file in order to publish. As a result, if the repository already has an .npmrc file, this will leave the file in a modified state. This isn't a big problem (it happens after creating a git commit, so the updated .npmrc file doesn't get published anywhere), but it has the potential to be confusing or cause human errors when generating a release locally. This commit updates the release publishing process to restore the old .npmrc file after the release finishes