Update: Major version changelogs include all prereleases
For a new stable release following a prerelease sequence, the changelog
will include the commits for all of the prereleases going back to the
last stable release. For example, given the following release sequence:

- 6.7.8
- 6.8.0
- 7.0.0-alpha.0
- 7.0.0-alpha.1
- 7.0.0

The `7.0.0` changelog will include all changes since 6.8.0, whereas the
`7.0.0-alpha.1` changelog will include only changes since
`7.0.0-alpha.0`.

If we prefer to include all changes going back to the previous stable
release even for prerelease changelogs, that's an easy change and
actually simplifies the logic a bit.